pyproject.toml
[project]
name = "ping"
version = "0.1.0"
description = "Sistema RAG para análise de posts do Instagram institucional"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=4.0.0",
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "ollama>=0.1.0",
    "python-dateutil>=2.8.0",
    "emoji>=2.0.0",
]
